zip_extract: handle directories within zip files
mz_zip_reader_extract_to_file fails when you ask it to extract a directory (according to the comment on line 621 of miniz.h and confirmed by my tests), thus zip_extract will fail if it attempts to decompress a zip file if that zip file contains a directory (there is an exception but I will not go into it here). This patch corrects the issue.