Added test for delete from in-memory zip.
This required making some updates that are kind of hacky.
The path of least edits to miniz necessitated updating miniz
mz_zip_reader_init_mem to initialize the writer function and
additionally zip_stream_open to handle the case where a user passes
a stream and the mode 'r' or 'd'.