Support zip archives not starting at zero offset
Zip archives are usually read from the back to the front, making it
possible to append them to other files (e.g. executables) in order to
provide some kind of embedded filesystem.

This PR makes it possible to read, write, and update such zip files
without affecting the existing functionality.